+++ debian/control
@@ -24,6 +24,7 @@ Depends: ${misc:Depends},
          python-chardet,
          python-cherrypy3,
          python-configobj,
+         python-configparser,
          python-cryptography,
          python-decorator,
          python-feedparser,
