+++ doc/development/development_on_windows.rst
@@ -172,7 +172,7 @@ There are some additional packages which should be installed. They can easily be
 .. code-block:: none
 
     pip install cython  # Needs to be installed first for meliae
-    pip install bitcoinlib cherrypy chardet configobj cryptography decorator feedparser meliae netifaces networkx pillow psutil twisted libnacl
+    pip install bitcoinlib cherrypy chardet configobj configparser cryptography decorator feedparser meliae netifaces networkx pillow psutil twisted libnacl
 
 Running Tribler
 ---------------
